requirements.txt
# Core dependencies
pytest==7.4.0
pytest-bdd==6.1.1
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.0
numpy==1.25.2

# AWS Bedrock integration
boto3==1.28.40
botocore==1.31.40

# RAGAS evaluation (minimal)
nltk>=3.8.1
scikit-learn>=1.3.0

# Testing and development
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
